backend/requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
flask-cors==5.0.0
requests==2.31.0
xarray==2024.3.0
netCDF4==1.6.5
h5netcdf==1.3.0
scipy==1.13.0
google-generativeai==0.8.6
python-dotenv==1.0.1

